requirements.txt
gunicorn==19.8.1
dj-database-url==0.5.0
Django==2.0.5
django-heroku==0.3.1
django-webpack-loader==0.6.0
djangorestframework==3.8.2
whitenoise==3.3.1